Redesign functions
Functions now take only a single argument,
and evaluate a single expression.
If you want to pass more than one argument, use a tuple.
There's even a special tuple-call syntax: F{2, 3}.
If you want to evaluate more than a single expression in a function,
you don't have to, since tuples (and lists) are eager and will evaluate
all elements in order!
And in order to retrieve the last element of such a tuple,
there is the <- operator,
which just matches the LHS with the last element of the RHS tuple.
If the RHS of the <- operator is function, a normal match occures,
but the function is wrapped in such a way that only the last element in
the tuple that it returns is actually returned.